Move eventList to REST.js and fix address_line in event preview
