+++ frontend/src/event/EventList.css
@@ -64,11 +64,22 @@
   display: flex;
   width: 100%;
   justify-content: space-between;
-  font-size: small;
 }
 
 .event-preview-title {
   font-size: large;
+}
+
+.event-preview-location {
+  font-size: xx-small;
+  width: 14em;
+  text-overflow: clip;
+}
+
+.event-preview-date {
+  flex-grow: 1;
+  text-align: right;
+  font-size: small;
 }
 
 @media screen and (min-width: 1081px) {
